- Checkpoint

+++ include/Sensors.hpp
@@ -43,7 +43,7 @@ namespace Sensors
     auto calibrateMagnetometer() -> bool;
 
     auto distances() -> std::array<std::pair<int, float>, 6>;
-    auto colors() -> std::array<uint16_t, 3>;
+    auto colors() -> std::array<float, 3>;
     auto rotation() -> std::array<float, 3>;
     auto acceleration() -> std::array<float, 3>;
     auto magnetic() -> std::array<float, 3>;
